General Category strings
This generates a directory of strings, one per file, containing each
Unicode general category, and a master file which imports them.  That
should combine nicely with lazy compilation, since the Cn category in
particular is absolutely enormous, is only there for completeness, and
therefore shouldn't even be opened unless someone is silly enough to
want to use it for something.

This approach seems to be a deviation from approved style, which would
use the build system more directly somehow.  I'm not convinced of any
disadvantages to doing it this way, it seems like doing otherwise would
be extra ceremony which this library would not benefit from.