Add General Category RuneSets
This seems to have worked well enough. I'm going to punt on testing it,
since the building blocks are themselves well tested, and just drudge my
way through some more Unicode files.

Next up, Scripts.  That promises to be only a small variation on General
Categories, with the wrinkle that I'll do both primary and extended sets
so it's not quite worth it to try and factor out the common logic into
a helper routine.  Some of the other derived properties might be close
enough to identical to justify that effort, we'll see.