Add Blocks.txt
In a way, this is the classic case for the RuneSet approach being outbid
by simpler, specialized code.  These are just ranges, and as such any of
them can be checked as any other range.

However, it's not clear that the decoding cost is less than the range
test in combination.  A range is certainly smaller.  Most relevantly for
our purposes, RuneSets have set operations, so these serve as a building
uhm.  Block.